ref: extract validation check for invalid chars into a separate function
