+++ src/app/modules/form/admin-form/admin-form.middlewares.ts
@@ -5,7 +5,8 @@ import {
   FormStatus,
   SettingsUpdateDto,
 } from '../../../../../shared/types'
-import { UNICODE_ESCAPED_REGEX } from '../form.utils'
+
+import { verifyValidUnicodeString } from './admin-form.utils'
 
 /**
  * Joi validator for PATCH /forms/:formId/settings route.
@@ -29,20 +30,5 @@ export const updateSettingsValidator = celebrate({
     }).min(1),
   })
     .min(1)
-    .custom((value, helpers) => {
-      // If there are invalid utf-8 encoded unicode-escaped characters,
-      // nodejs treats the sequence of characters as a string e.g. \udbbb is treated as a 6-character string instead of an escaped unicode sequence
-      // If this is saved into the db, an error is thrown when the driver attempts to read the db document as the driver interprets this as an escaped unicode sequence.
-      // Since valid unicode-escaped characters will be processed correctly (e.g. \u00ae is processed as ®), they will not trigger an error
-      // Also note that if the user intends to input a 6-character string of the same form e.g. \udbbb, the backslash will be escaped (i.e. double backslash) and hence this will also not trigger an error
-
-      const valueStr = JSON.stringify(value)
-
-      if (UNICODE_ESCAPED_REGEX.test(valueStr)) {
-        return helpers.message({
-          custom: 'There are invalid characters in your input',
-        })
-      }
-      return value
-    }),
+    .custom((value, helpers) => verifyValidUnicodeString(value, helpers)),
 })
