+++ src/app/modules/form/admin-form/admin-form.utils.ts
@@ -40,6 +40,7 @@ import {
   PrivateFormError,
   TransferOwnershipError,
 } from '../form.errors'
+import { UNICODE_ESCAPED_REGEX } from '../form.utils'
 
 import {
   CreatePresignedUrlError,
@@ -444,4 +445,27 @@ export const generateTwilioCredSecretKeyName = (formId: string): string =>
 export const checkIsApiSecretKeyName = (msgSrvcName: string): boolean => {
   const prefix = `formsg/${config.secretEnv}/api`
   return msgSrvcName.startsWith(prefix)
+}
+
+/**
+ * Validation check for invalid utf-8 encoded unicode-escaped characteres
+ * @param value
+ * @param helpers
+ * @returns custom err message if there are invalid characters in the input
+ */
+export const verifyValidUnicodeString = (value: any, helpers: any) => {
+  // If there are invalid utf-8 encoded unicode-escaped characters,
+  // nodejs treats the sequence of characters as a string e.g. \udbbb is treated as a 6-character string instead of an escaped unicode sequence
+  // If this is saved into the db, an error is thrown when the driver attempts to read the db document as the driver interprets this as an escaped unicode sequence.
+  // Since valid unicode-escaped characters will be processed correctly (e.g. \u00ae is processed as ®), they will not trigger an error
+  // Also note that if the user intends to input a 6-character string of the same form e.g. \udbbb, the backslash will be escaped (i.e. double backslash) and hence this will also not trigger an error
+
+  const valueStr = JSON.stringify(value)
+
+  if (UNICODE_ESCAPED_REGEX.test(valueStr)) {
+    return helpers.message({
+      custom: 'There are invalid characters in your input',
+    })
+  }
+  return value
 }
